CNN-based image classification model.
This project implements a CNN model to classify images from the CIFAR-100 dataset. It includes data preprocessing, model training, and performance evaluation. The images are resized from 32x32 to 64x64 resolution for better feature extraction. The model is trained using TensorFlow/Keras and achieves a training accuracy of 84% and a validation accuracy of 67%.